pyproject.toml
[build-system]
requires = [
  "setuptools>=61.0",
  "certifi>=14.05.14",
  "typer>=0.16",
  "inquirer>=3.4.0",
  "recodex_cli_lib_eceltov>=0.0.63"
]
build-backend = "setuptools.build_meta"

[project]
name = "recodex_cli_eceltov"
version = "0.0.18"
authors = [
  { name="Vojtech Kloda", email="vojtech.kloda@gmail.com" },
]
description = "A module for ReCodEx API calls."
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
  "setuptools>=61.0",
  "certifi>=14.05.14",
  "typer>=0.16",
  "inquirer>=3.4.0",
  "recodex_cli_lib_eceltov>=0.0.63"
]

[project.urls]
Homepage = "https://github.com/ReCodEx"
Issues = "https://github.com/ReCodEx"
